+++ Source/Core/Visualization/Widget/ColorMapBar.cpp
@@ -105,7 +105,7 @@ void ColorMapBar::paintEvent()
 {
     this->screenUpdated();
 
-    if ( !BaseClass::isShown() ) return;
+    if ( !BaseClass::isVisible() ) return;
 
     if ( !m_texture_downloaded )
     {
@@ -201,6 +201,9 @@ void ColorMapBar::resizeEvent( int width, int height )
 {
     kvs::IgnoreUnusedVariable( width );
     kvs::IgnoreUnusedVariable( height );
+    const auto p = BaseClass::anchorPosition();
+    Rectangle::setPosition( p.x(), p.y() );
+
     this->screenResized();
 }