+++ Source/Core/Visualization/Widget/Label.cpp
@@ -3,14 +3,6 @@
  *  @file   Label.cpp
  *  @author Naohisa Sakamoto
  */
-/*----------------------------------------------------------------------------
- *
- *  Copyright (c) Visualization Laboratory, Kyoto University.
- *  All rights reserved.
- *  See http://www.viz.media.kyoto-u.ac.jp/kvs/copyright/ for details.
- *
- *  $Id$
- */
 /*****************************************************************************/
 #include "Label.h"
 #include <kvs/OpenGL>
@@ -87,9 +79,9 @@ int Label::adjustedWidth()
     size_t max_width = 0;
     BaseClass::painter().begin( BaseClass::screen() );
     const kvs::FontMetrics metrics = BaseClass::painter().fontMetrics();
-    for ( size_t i = 0; i < m_text.size(); i++ )
+    for ( size_t i = 0; i < m_text_list.size(); i++ )
     {
-        const size_t line_width = metrics.width( m_text[i] );
+        const size_t line_width = metrics.width( m_text_list[i] );
         max_width = kvs::Math::Max( max_width, line_width );
     }
     BaseClass::painter().end();
@@ -107,7 +99,7 @@ int Label::adjustedHeight()
 {
     BaseClass::painter().begin( BaseClass::screen() );
     const kvs::FontMetrics metrics = BaseClass::painter().fontMetrics();
-    const size_t nlines = m_text.size();
+    const size_t nlines = m_text_list.size();
     const size_t character_height = metrics.height();
     BaseClass::painter().end();
     return nlines * character_height + BaseClass::margin() * 2;
@@ -122,7 +114,7 @@ void Label::paintEvent()
 {
     this->screenUpdated();
 
-    if ( !BaseClass::isShown() ) return;
+    if ( !BaseClass::isVisible() ) return;
 
     BaseClass::painter().begin( BaseClass::screen() );
     BaseClass::drawBackground();
@@ -130,10 +122,10 @@ void Label::paintEvent()
     const int x = BaseClass::x() + BaseClass::margin();
     const int y = BaseClass::y() + BaseClass::margin();
     const size_t character_height = BaseClass::painter().fontMetrics().height();
-    for ( size_t line = 0; line < m_text.size(); line++ )
+    for ( size_t line = 0; line < m_text_list.size(); line++ )
     {
         const kvs::Vec2i p( x, y + character_height * ( line + 1 ) );
-        BaseClass::painter().drawText( p, m_text[line] );
+        BaseClass::painter().drawText( p, m_text_list[line] );
     }
 
     BaseClass::painter().end();
@@ -150,6 +142,8 @@ void Label::resizeEvent( int width, int height )
 {
     kvs::IgnoreUnusedVariable( width );
     kvs::IgnoreUnusedVariable( height );
+    const auto p = BaseClass::anchorPosition();
+    Rectangle::setPosition( p.x(), p.y() );
     this->screenResized();
 }